Create a BukuImporter class
